remove babel-eslint parser in favor of eslint native parser
Babel-eslint parser github page says that for ECMA 2015/2016/2017/JSX
the native parser is feature complete.
Added to it, babel-eslint seems to be receiving lower support.

I don't remember why we needed to use a third party parser, but it's
properly working with the native one, so let's switch back to it.